not include payment

+++ FahasaStoreAPI/Models/DimensionModel.cs
@@ -3,9 +3,9 @@
     public class DimensionModel
     {
         public int Id { get; set; }
-        public double Length { get; set; }
-        public double Width { get; set; }
-        public double Height { get; set; }
+        public int Length { get; set; }
+        public int Width { get; set; }
+        public int Height { get; set; }
         public string Unit { get; set; } = null!;
         public DateTime? CreatedAt { get; set; }
     }
